requirements.txt
opencv-python-headless>=4.9.0
numpy>=2.0.0
pillow>=10.0.0
requests>=2.31.0
google-genai>=1.0.0
